Create a new VS 2010 solution for LunarGLASS that modularizes the core, allowing different front ends to be added as separate projects.
This now lives in a new Build directory.  The old solution is still there in standalone, but perhaps it will eventually be removed, depending on the utility of the new modular front end method (possibly one stays as a VS 2010 Express build and one as a VS 2010 build).

The current glsl2 front end is now a project in the new solution that builds the same thing the existing standalone built.  But, it names it glsl2.exe instead of standalone.exe.  A future front end can then build an executable with a different name.

Back ends are still switched at runtime, internally.  For front ends, this is more difficult, as they might have their .l and .y file (and main()).  However, to keep mesa dependencies outside of the LunarGLASS core, I turned off the existing partial TGSI back end, which is quite out of date now.